pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gauge"]

[project]
name = "libgauge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pillow>=12.1.1",
]

[dependency-groups]
dev = [
    "basedpyright>=1.38.2",
    "mypy>=1.0",
    "orjson>=3.11.7",
    "pytest>=9.0.2",
    "ruff>=0.15.6",
    "syrupy>=5.1.0",
]


[tool.ruff]
target-version = "py314"
line-length = 88

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "RUF", # ruff-specific rules
]

[tool.ruff.lint.isort]
known-first-party = ["gauge"]

[tool.basedpyright]
typeCheckingMode = "all"

[tool.mypy]
strict = true
